edit

+++ src/main/java/kg/attractor/job_search_java_25/repository/WorkExperienceInfoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkExperienceInfoRepository extends JpaRepository<WorkExperienceInfo, Integer> {
+public interface WorkExperienceInfoRepository extends JpaRepository<WorkExperienceInfo, Long> {
     void deleteWorkExperienceInfoById(Long id);
 }
